+++ changelog.md
@@ -5,6 +5,10 @@
 ### Other
 - Updated some spinner values in the add publisher window.
 
+### Bug fixes
+- Fixed export fail when a npc engine was exported where genre/platform ids where ``-1``
+  - This would only occur when the safety features where off, and it was tried to export the ``without engine`` npc engine
+
 ## [v4.0.0](https://github.com/LMH01/MGT2_Mod_Tool/releases/tag/v4.0.0) (Latest Version)
 
 ### Important